requirements.txt
# Core
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0
imageio>=2.9.0

# Experiment Tracking (optional but recommended)
wandb>=0.15.0
pyyaml>=6.0

# Optional (for development)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Jupyter (optional)
jupyter>=1.0.0
ipython>=8.0.0
